+++ CHANGELOG.md
@@ -66,6 +66,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - The caret is now only displayed when an editor is on focus (previously it was
   always displayed).
 
+- The `<code-sample>` side-by-side layout is now consistently 70%/30% (widths be
+  changed using the `editor` and `preview` CSS shadow parts).
+
 ### Added
 
 - Add syntax highlighting of TypeScript files.
@@ -107,6 +110,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - Fix absent CSS syntax highlighting.
+- Various styling/layout glitches.
 
 ## [0.1.0-pre.3] - 2020-10-05